Refractor `LiteReceiverConnectionToken`
